refactor(black-box): isolate Hetzner parity recipe
